+++ app_fenrir/src/main/kotlin/dev/ragnarok/fenrir/domain/IPollInteractor.kt
@@ -14,15 +14,15 @@ interface IPollInteractor {
         options: List<String>
     ): Single<Poll>
 
-    fun addVote(accountId: Int, poll: Poll, answerIds: Set<Int>): Single<Poll>
-    fun removeVote(accountId: Int, poll: Poll, answerId: Int): Single<Poll>
+    fun addVote(accountId: Int, poll: Poll, answerIds: Set<Long>): Single<Poll>
+    fun removeVote(accountId: Int, poll: Poll, answerId: Long): Single<Poll>
     fun getPollById(accountId: Int, ownerId: Int, pollId: Int, isBoard: Boolean): Single<Poll>
     fun getVoters(
         accountId: Int,
         ownerId: Int,
         pollId: Int,
         isBoard: Int?,
-        answer_ids: List<Int>,
+        answer_ids: List<Long>,
         offset: Int?,
         count: Int?
     ): Single<List<User>>
